[DEV-7191] Addresses PR feedback
